Use dd_run_check for ecs_fargate (#10374)
* Use dd_run_check

* Use dd_run_check for tests in ecs_fargate and update conftest

* Update style

* Update conftest

* Update ecs_fargate/tests/conftest.py


* Remove exchange_server changes

* Remove unused import
